+++ src/__tests__/utilities/form/fields/componentBuilders.test.js
@@ -90,17 +90,17 @@ describe('Dynamic Fields componentBuilders', () => {
               enum: ['1', '2'],
               userInputAllowed: true
             })
-          ).toEqual({ component: AutoComplete, props: { enums: ['1', '2'], fixParentNode: true } })
+          ).toEqual({ component: AutoComplete, props: { enums: ['1', '2'] } })
         })
       })
 
       describe('without userInputAllowed', () => {
-        it('returns a Select', () => {
+        it('returns a searchable Select', () => {
           expect(
             buildComponentForString({
               enum: ['1', '2']
             })
-          ).toEqual({ component: Select, props: { enums: ['1', '2'], fixParentNode: true } })
+          ).toEqual({ component: Select, props: { enums: ['1', '2'], fixParentNode: true, showSearch: true } })
         })
       })
     })
@@ -122,15 +122,14 @@ describe('Dynamic Fields componentBuilders', () => {
               valueEnum: [
                 ['One', '1'],
                 ['Two', '2']
-              ],
-              fixParentNode: true
+              ]
             }
           })
         })
       })
 
       describe('without userInputAllowed', () => {
-        it('returns a Select', () => {
+        it('returns a searchable Select', () => {
           expect(
             buildComponentForString({
               valueEnum: [
@@ -145,7 +144,8 @@ describe('Dynamic Fields componentBuilders', () => {
                 ['One', '1'],
                 ['Two', '2']
               ],
-              fixParentNode: true
+              fixParentNode: true,
+              showSearch: true
             }
           })
         })