+++ src/__tests__/utilities/form/fields/index.test.js
@@ -1,5 +1,5 @@
 import { Input, Select, Switch } from '~su/components'
-import buildFields, { testExports } from '~su/utilities/form/fields'
+import buildFields, { normalizeEmptyToUndef, testExports } from '~su/utilities/form/fields'
 
 describe('Dynamic Fields building', () => {
   describe('buildFields', () => {
@@ -14,7 +14,7 @@ describe('Dynamic Fields building', () => {
     const extraConfigs = {}
     const namePrefix = null
 
-    const t = jest.fn().mockImplementation(() => { help: 'test' })
+    const t = jest.fn().mockImplementation((key) => { hint: `${key}_test` })
 
     it('builds fields using the schema', () => {
       expect(buildFields(schema, extraConfigs, namePrefix, t)).toEqual([
@@ -22,10 +22,11 @@ describe('Dynamic Fields building', () => {
           item: {
             name: 'field_one',
             label: 'Field one',
-            rules: [{ required: false }, { type: 'string' }, expect.any(Function)],
+            rules: [{ required: false }, { type: 'string' }, expect.any(Function), { whitespace: true }],
             dependencies: [],
             hasFeedback: true,
-            valuePropName: 'value'
+            valuePropName: 'value',
+            normalize: normalizeEmptyToUndef
           },
           component: <Input />
         },
@@ -33,13 +34,13 @@ describe('Dynamic Fields building', () => {
           item: {
             name: 'field_two',
             label: 'Field two',
-            rules: [{ required: true }, { type: 'string' }, expect.any(Function)],
+            rules: [{ required: true }, { type: 'string' }, expect.any(Function), { whitespace: true }],
             dependencies: [],
             hasFeedback: true,
             valuePropName: 'value',
             initialValue: 'two'
           },
-          component: <Select enums={['one', 'two']} fixParentNode={true} />
+          component: <Select enums={['one', 'two']} fixParentNode={true} showSearch={true} />
         }
       ])
     })
@@ -63,10 +64,11 @@ describe('Dynamic Fields building', () => {
             item: {
               name: 'field_one',
               label: 'Field one',
-              rules: [{ required: false }, { type: 'string' }, expect.any(Function)],
+              rules: [{ required: false }, { type: 'string' }, expect.any(Function), { whitespace: true }],
               dependencies: [],
               hasFeedback: true,
               valuePropName: 'value',
+              normalize: normalizeEmptyToUndef,
               width: 2
             },
             component: <Input />
@@ -75,7 +77,7 @@ describe('Dynamic Fields building', () => {
             item: {
               name: 'field_two',
               label: 'Field two',
-              rules: [{ required: true }, { type: 'string' }, expect.any(Function)],
+              rules: [{ required: true }, { type: 'string' }, expect.any(Function), { whitespace: true }],
               dependencies: [],
               hasFeedback: true,
               valuePropName: 'value',
@@ -97,10 +99,11 @@ describe('Dynamic Fields building', () => {
             item: {
               name: 'field_one',
               label: 'Field one',
-              rules: [{ required: false }, { type: 'string' }, expect.any(Function)],
+              rules: [{ required: false }, { type: 'string' }, expect.any(Function), { whitespace: true }],
               dependencies: [],
               hasFeedback: true,
               valuePropName: 'value',
+              normalize: normalizeEmptyToUndef,
               width: 2
             },
             component: <Input allowClear />
@@ -132,10 +135,11 @@ describe('Dynamic Fields building', () => {
             item: {
               name: 'field_one',
               label: 'Field one',
-              rules: [{ required: false }, { type: 'string' }, expect.any(Function)],
+              rules: [{ required: false }, { type: 'string' }, expect.any(Function), { whitespace: true }],
               dependencies: [],
               hasFeedback: true,
-              valuePropName: 'value'
+              valuePropName: 'value',
+              normalize: normalizeEmptyToUndef
             },
             component: <Input />
           },
@@ -143,12 +147,12 @@ describe('Dynamic Fields building', () => {
             item: {
               name: 'field_two',
               label: 'Field two',
-              rules: [{ required: true }, { type: 'string' }, expect.any(Function)],
+              rules: [{ required: true }, { type: 'string' }, expect.any(Function), { whitespace: true }],
               dependencies: [],
               hasFeedback: true,
               valuePropName: 'value'
             },
-            component: <Select enums={['one', 'two']} fixParentNode={true} />
+            component: <Select enums={['one', 'two']} fixParentNode={true} showSearch={true} />
           },
           {
             item: {
@@ -168,7 +172,8 @@ describe('Dynamic Fields building', () => {
               rules: [{ required: false }, { type: 'boolean' }, expect.any(Function)],
               dependencies: [],
               hasFeedback: true,
-              valuePropName: 'checked'
+              valuePropName: 'checked',
+              normalize: normalizeEmptyToUndef
             },
             component: <Switch />
           }
@@ -216,7 +221,8 @@ describe('Dynamic Fields building', () => {
                   rules: [{ required: false }, { type: 'boolean' }, expect.any(Function)],
                   dependencies: [],
                   hasFeedback: true,
-                  valuePropName: 'checked'
+                  valuePropName: 'checked',
+                  normalize: normalizeEmptyToUndef
                 },
                 component: <Switch size="small" />
               }
@@ -261,10 +267,11 @@ describe('Dynamic Fields building', () => {
               item: {
                 name: 'field_one',
                 label: 'Field one',
-                rules: [{ required: false }, { type: 'string' }, expect.any(Function)],
+                rules: [{ required: false }, { type: 'string' }, expect.any(Function), { whitespace: true }],
                 dependencies: [],
                 hasFeedback: true,
-                valuePropName: 'value'
+                valuePropName: 'value',
+                normalize: normalizeEmptyToUndef
               },
               component: <Input />
             },
@@ -272,12 +279,12 @@ describe('Dynamic Fields building', () => {
               item: {
                 name: 'field_two',
                 label: 'Field two',
-                rules: [{ required: true }, { type: 'string' }, expect.any(Function)],
+                rules: [{ required: true }, { type: 'string' }, expect.any(Function), { whitespace: true }],
                 dependencies: [],
                 hasFeedback: true,
                 valuePropName: 'value'
               },
-              component: <Select enums={['one', 'two']} fixParentNode={true} />
+              component: <Select enums={['one', 'two']} fixParentNode={true} showSearch={true} />
             },
             {
               item: {
@@ -297,7 +304,8 @@ describe('Dynamic Fields building', () => {
                 rules: [{ required: false }, { type: 'boolean' }, expect.any(Function)],
                 dependencies: [],
                 hasFeedback: true,
-                valuePropName: 'checked'
+                valuePropName: 'checked',
+                normalize: normalizeEmptyToUndef
               },
               component: <Switch />
             }
@@ -346,7 +354,7 @@ describe('Dynamic Fields building', () => {
     const { buildFieldConfig } = testExports
 
     const name = 'field_one'
-    const properties = { type: 'string', minLength: 1 }
+    const properties = { type: 'string' }
     const required = true
     const conditionalRules = null
     const extraConfig = {}