+++ src/lib/components/Layout/index.jsx
@@ -12,22 +12,35 @@ import { StyledContent, StyledLayout } from './index.styled'
 
 const { useToken } = theme
 
-const Layout = ({ menuProps, sidebarItems, children, onSideMenuSelect, themeOverrides = {}, pathname }) => {
+const Layout = ({
+  menuProps,
+  sidebarItems,
+  children,
+  onSideMenuSelect,
+  themeOverrides = {},
+  pathname,
+  sideMenuChildren,
+  contentContainerStyles
+}) => {
   const { token: themeToken } = useToken()
 
   return (
     <StyledLayout>
       <TopNav themeOverrides={themeOverrides} {...menuProps} />
       <AntdLayout>
-        <SideMenu
-          sidebarItems={sidebarItems}
-          onSideMenuSelect={onSideMenuSelect}
-          themeToken={themeToken}
-          pathname={pathname}
-        />
+        {!!sidebarItems.length && (
+          <SideMenu
+            sidebarItems={sidebarItems}
+            onSideMenuSelect={onSideMenuSelect}
+            themeToken={themeToken}
+            pathname={pathname}
+          >
+            {sideMenuChildren}
+          </SideMenu>
+        )}
         <AntdLayout>
           <GlobalAlert />
-          <StyledContent>{children}</StyledContent>
+          <StyledContent $styleOptions={contentContainerStyles}>{children}</StyledContent>
         </AntdLayout>
       </AntdLayout>
     </StyledLayout>
@@ -40,7 +53,9 @@ Layout.propTypes = {
   onSideMenuSelect: SideMenu.propTypes.onSideMenuSelect,
   children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.node), PropTypes.node]),
   themeOverrides: PropTypes.object,
-  pathname: PropTypes.string
+  pathname: PropTypes.string,
+  sideMenuChildren: PropTypes.node,
+  contentContainerStyles: PropTypes.string
 }
 
 export default withLoader(Layout)