+++ src/lib/constants/componentsConfig.js
@@ -2,7 +2,10 @@ import { COLORS } from './index'
 
 const components = {
   Menu: {
-    itemDisabledColor: COLORS.white
+    itemDisabledColor: COLORS.white,
+    itemSelectedColor: COLORS.white,
+    itemHoverColor: COLORS.white,
+    itemHoverBg: 'transparent'
   }
 }